% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms.R
\name{cms_home_health}
\alias{cms_home_health}
\title{Home health agency information}
\usage{
cms_home_health(state = NULL, max_results = 1000)
}
\arguments{
\item{state}{2-letter state code}

\item{max_results}{Max results (default 1000)}
}
\value{
tibble with agency info and quality ratings
}
\description{
Home health agency information
}
